Fixed I/O tooltip
It was populated by senseless numbers because two variables weren't initialized to 0.

The patch fixes the issue and shows processes only when there is real I/O.

Fixes https://github.com/lxqt/qps/issues/408

+++ src/misc.h
@@ -57,7 +57,7 @@ void setQpsTheme();
 int fsize(char *fname);
 void msleep(long msec);
 void mem_string(int kbytes, char *buf);
-void mem_string_k(int kbytes, char *buf);
+void mem_string_k(unsigned long kbytes, char *buf);
 
 void init_xpm();
 void init_misc(QWidget *main);
